Refactor cyral_repository
