+++ cyral/provider/provider.go
@@ -21,7 +21,6 @@ import (
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/policy"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/policy/rule"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/regopolicy"
-	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/repository"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/repository/accessgateway"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/repository/accessrules"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/repository/binding"
@@ -112,7 +111,6 @@ func getDataSourceMap(ps []core.PackageSchema) map[string]*schema.Resource {
 	schemaMap["cyral_integration_idp_saml"] = idpsaml.DataSourceIntegrationIdPSAML()
 	schemaMap["cyral_integration_logging"] = logging.DataSourceIntegrationLogging()
 	schemaMap["cyral_permission"] = permission.DataSourcePermission()
-	schemaMap["cyral_repository"] = repository.DataSourceRepository()
 	schemaMap["cyral_role"] = role.DataSourceRole()
 	schemaMap["cyral_saml_configuration"] = samlconfiguration.DataSourceSAMLConfiguration()
 	schemaMap["cyral_sidecar_bound_ports"] = sidecar.DataSourceSidecarBoundPorts()
@@ -168,7 +166,6 @@ func getResourceMap(ps []core.PackageSchema) map[string]*schema.Resource {
 	schemaMap["cyral_policy"] = policy.ResourcePolicy()
 	schemaMap["cyral_policy_rule"] = rule.ResourcePolicyRule()
 	schemaMap["cyral_rego_policy_instance"] = regopolicy.ResourceRegoPolicyInstance()
-	schemaMap["cyral_repository"] = repository.ResourceRepository()
 	schemaMap["cyral_repository_access_rules"] = accessrules.ResourceRepositoryAccessRules()
 	schemaMap["cyral_repository_access_gateway"] = accessgateway.ResourceRepositoryAccessGateway()
 	schemaMap["cyral_repository_binding"] = binding.ResourceRepositoryBinding()
