+++ cyral/provider/schema_loader.go
@@ -6,6 +6,7 @@ import (
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/integration/hcvault"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/integration/slack"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/integration/teams"
+	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/repository"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/repository/datamap"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/repository/useraccount"
 	"github.com/cyralinc/terraform-provider-cyral/cyral/internal/samlcertificate"
@@ -17,6 +18,7 @@ func packagesSchemas() []core.PackageSchema {
 		datalabel.PackageSchema(),
 		datamap.PackageSchema(),
 		hcvault.PackageSchema(),
+		repository.PackageSchema(),
 		samlcertificate.PackageSchema(),
 		slack.PackageSchema(),
 		teams.PackageSchema(),
